+++ src/audiodevice.rs
@@ -508,6 +508,8 @@ pub fn get_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             silence_timeout: conf.silence_timeout,
             skip_bytes,
             read_bytes,
+            stop_on_rate_change: conf.stop_on_rate_change,
+            rate_measure_interval: conf.rate_measure_interval
         }),
         config::CaptureDevice::Stdin {
             channels,
@@ -529,6 +531,8 @@ pub fn get_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             silence_timeout: conf.silence_timeout,
             skip_bytes,
             read_bytes,
+            stop_on_rate_change: conf.stop_on_rate_change,
+            rate_measure_interval: conf.rate_measure_interval
         }),
         #[cfg(all(feature = "cpal-backend", target_os = "macos"))]
         config::CaptureDevice::CoreAudio {
@@ -568,6 +572,8 @@ pub fn get_capture_device(conf: config::Devices) -> Box<dyn CaptureDevice> {
             sample_format: format,
             silence_threshold: conf.silence_threshold,
             silence_timeout: conf.silence_timeout,
+            stop_on_rate_change: conf.stop_on_rate_change,
+            rate_measure_interval: conf.rate_measure_interval
         }),
         #[cfg(all(feature = "cpal-backend", feature = "jack-backend"))]
         config::CaptureDevice::Jack { channels, device } => {